balanced power up spawn rates

+++ ZoeSlasher/Game/Spawners/PotionSpawner.swift
@@ -27,14 +27,14 @@ class PotionSpawner {
         trySpawnPotion(type: .health, amount: 10, timer: &timeSinceLastHealthPotion, interval: healthPotionInterval)
     }
     
-    func spawnPotion(type: PotionType, amount: Int = 0) {
+    func spawnPotion(type: PotionType, amount: Float = 0) {
         let potion = Potion(type: type, amount: amount)
         potion.position = scene.randomPosition(padding: [300, 200])
         potion.parent = scene.rootNode
         scene.potions.insert(potion)
     }
     
-    private func trySpawnPotion(type: PotionType, amount: Int, timer: inout TimeInterval, interval: TimeInterval) {
+    private func trySpawnPotion(type: PotionType, amount: Float, timer: inout TimeInterval, interval: TimeInterval) {
         if timer >= interval {
             spawnPotion(type: type, amount: amount)
